Azure/Azure-export-alerts/Remote_execution/requirements.txt
xlwt
pandas
datetime
openpyxl
azure.identity
azure.mgmt.monitor 
azure.mgmt.subscription 